requirements.txt
# requirements.txt - Working version
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
statsmodels>=0.14.0
scikit-learn>=1.3.0
yfinance>=0.2.18
sqlalchemy>=2.0.0
xlsxwriter>=3.1.0
openpyxl>=3.1.0
fpdf2>=2.7.0
jupyter>=1.0.0
ipython>=8.14.0
pytest>=7.3.0
python-dateutil>=2.8.0
pytz>=2023.0
requests>=2.28.0